Move tile and gradient utilities
